content-bot/requirements.txt
openai==1.10.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
flask==3.0.0
flask-cors==4.0.0
anthropic==0.18.0
rich==13.7.0
